Combine all compat public types into index.d.ts
